edited Chinese agricultural investment paragraph
In China’s Belt and Road Initiative a national food security strategy was defined for China to utilize labor resources to develop foreign land, water, and energy resources. Rice, grain, and rubber were the most reported planted crops between 2010 and 2017. Cameroon alone, accounted for 41% of all the land actually acquired by the Chinese (SAIS-CARI). CARI Director Deborah Brautigam’s writes if the “Chinese government is aggressively buying huge tracts of prime African land to grow food to ship back to China” Will Africa Feed China? in the future? Africa is a developing continent with an abundance of arable farmland and Chinese investors CITIC Construction and CAMC Engineering were the largest recorded investor in Angola between 2011 and 2014. Specifically GMG Global/Sinochem acquired the most land in Africa to grow rubber. China recognizes the advantages to investing in agricultural projects in Africa and it is our responsibility to be aware of the reasons.